# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VulnOps Hub
  description: Vulnerability-operations control plane — modular monolith
  version: 0.1.0
paths:
...
  /api/v1/organizations/{org_id}/cases:
    post:
      tags:
      - cases
      summary: Create Case
      operationId: create_case_api_v1_organizations__org_id__cases_post
      security:
      - BearerAuth: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseCreateResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              additionalProperties: false
              description: Canonical, actor-free case creation request contract.
              properties:
                title:
                  minLength: 1
                  title: Title
                  type: string
                owner_team:
                  minLength: 1
                  title: Owner Team
                  type: string
                priority:
                  enum:
                  - P0
                  - P1
                  - P2
                  - P3
                  - P4
                  title: Priority
                  type: string
                exposures:
                  items:
                    type: string
                  title: Exposures
                  type: array
                policy_version:
                  anyOf:
                  - minLength: 1
                    type: string
                  - type: 'null'
                  title: Policy Version
                assignee:
                  anyOf:
                  - minLength: 1
                    type: string
                  - type: 'null'
                  title: Assignee
              required:
              - title
              - owner_team
              - priority
              title: CaseCreateRequest
              type: object
    get:
      tags:
      - cases
      summary: List Cases
      operationId: list_cases_api_v1_organizations__org_id__cases_get
      security:
      - BearerAuth: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: priority
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
      - name: owner_team
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Team
      - name: assignee
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignee
      - name: sla_breached
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Sla Breached
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          type: string
          pattern: ^-?(created_at|updated_at|due_at|priority)$
          default: -created_at
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseListResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/organizations/{org_id}/cases/{case_id}:
    get:
      tags:
      - cases
      summary: Get Case
      operationId: get_case_api_v1_organizations__org_id__cases__case_id__get
      security:
      - BearerAuth: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDetailResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/organizations/{org_id}/cases/{case_id}/allowed-transitions:
    get:
      tags:
      - cases
      summary: Allowed Transitions
      operationId: allowed_transitions_api_v1_organizations__org_id__cases__case_id__allowed_transitions_get
      security:
      - BearerAuth: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowedTransitionsResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/organizations/{org_id}/cases/{case_id}/transitions:
    post:
      tags:
      - cases
      summary: Transition Case
      operationId: transition_case_api_v1_organizations__org_id__cases__case_id__transitions_post
      security:
      - BearerAuth: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransitionResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              additionalProperties: false
              description: Actor-free case transition request contract.
              properties:
                target:
                  minLength: 1
                  title: Target
                  type: string
                reason:
                  anyOf:
                  - type: string
                  - type: 'null'
                  title: Reason
              required:
              - target
              title: TransitionRequest
              type: object
...
components:
  schemas:
...
    CaseCreateResponse:
      properties:
        id:
          type: string
          title: Id
        case_key:
          type: string
          title: Case Key
        title:
          type: string
          title: Title
        status:
          type: string
          title: Status
        priority:
          type: string
          title: Priority
        owner_team:
          type: string
          title: Owner Team
        organization_id:
          type: string
          title: Organization Id
        version:
          type: integer
          title: Version
        etag:
          type: string
          title: Etag
        due_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Due At
        ownership_escalated:
          type: boolean
          title: Ownership Escalated
          default: false
      type: object
      required:
      - id
      - case_key
      - title
      - status
      - priority
      - owner_team
      - organization_id
      - version
      - etag
      title: CaseCreateResponse
...
    CaseListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CaseDetailResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: CaseListResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ProblemDetails:
      properties:
        type:
          type: string
          title: Type
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        code:
          type: string
          title: Code
        detail:
          type: string
          title: Detail
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
        fields:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Fields
      type: object
      required:
      - type
      - title
      - status
      - code
      - detail
      title: ProblemDetails
      description: Stable error envelope shared by authenticated API operations.
...
    TransitionResponse:
      properties:
        id:
          type: string
          title: Id
        status:
          type: string
          title: Status
        version:
          type: integer
          title: Version
        etag:
          type: string
          title: Etag
      type: object
      required:
      - id
      - status
      - version
      - etag
      title: TransitionResponse
...
  securitySchemes:
    BearerAuth:
      type: http
      description: OIDC access token
      scheme: bearer
      bearerFormat: JWT